prepare for Epic Super Bowser World

+++ script.d/common.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 
 TMD_OPTS="--tt-id=K"
-TMP_FILES=(Another Another_Super_Mario_Brothers_Wii_2.0.zip nsmb.d Newer_Super_Mario_Bros._Wii_HS.zip XmasNewer NewerSMBW.zip NewerFiles "Newer Summer Sun" Newer_Summer_Sun.zip ZPW_1.1.ips lozpw110.rar "*[Ll]ink*[Pp]ast*smc")
+TMP_FILES=(Another Another_Super_Mario_Brothers_Wii_2.0.zip nsmb.d Newer_Super_Mario_Bros._Wii_HS.zip XmasNewer NewerSMBW.zip NewerFiles "Newer Summer Sun" Newer_Summer_Sun.zip ZPW_1.1.ips lozpw110.rar "*[Ll]ink*[Pp]ast*smc" Epic_Super_Bowser_World_v1.00 Epic_Super_Bowser_World_v1.00.zip)
+DL_BANNER=FALSE
 
 setup_tools () {
 
@@ -45,8 +46,10 @@ A	NewerSMB
 B	Newer Summer Sun
 C	AnotherSMB
 D	Newer: Holiday Special
+XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX (coming soon)
+F	Epic Super Bowser World
 
-F	The Legend of Zelda: Parallel Worlds
+G	The Legend of Zelda: Parallel Worlds
 "
 
 read GAME
@@ -82,7 +85,7 @@ download_banner () {
 
 apply_banner () {
 
-	if [[ -e ${BANNER} ]]; then
+	if [[ ${BANNER} && -e ${BANNER} ]]; then
 		cp "${BANNER}" "${BANNER_LOCATION}"
 	else
 		echo "specified banner ${BANNER} does not exist, not modifying"
